Rename ec-core::generation to ec-core::distributions
